Extract registries as structures (collections of handlers)
