+++ kopf/structs/registries.py
