+++ tests/registries/test_global_registry.py
@@ -1,6 +1,6 @@
 import collections
 
-from kopf.reactor.registries import GlobalRegistry
+from kopf.structs.registries import GlobalRegistry
 from kopf.structs.resources import Resource
